requirements.txt
numpy>=1.26.4
pandas>=2.2.2
pillow>=10.4.0
roboflow>=1.1.44
streamlit>=1.38.0
ultralytics>=8.2.91